+++ src/Synapse/Synapse/ChangeLog.md
@@ -17,7 +17,11 @@
     * Overview of change #1
         - Additional information about change #1
 -->
+
 ## Upcoming Release
+* Added cmdlets for Synapse Integration Runtime
+	- Added `Start-AzSynapseIntegrationRuntime` cmdlet
+	- Added `Stop-AzSynapseIntegrationRuntime` cmdlet
 * Added cmdlets for Synapse trigger run
 	- Added `Stop-AzSynapseTriggerRun` cmdlet
 	- Added `Invoke-AzSynapseTriggerRun` cmdlet